fmt: Clean REST API Testing
